Refactor Message Form to its own widget
